fix(ezpz): default `training.dtype` to float32 so RMSNorm weights actually train
With bf16 master weights (the previous default), RMSNorm.weight init
1.0 has bf16 ulp 7.8e-3, but the per-step optimizer update is
~1.6e-5 — every update rounds to zero forever. All RMSNorm
parameters (`norm.weight`, `layers.*.attention_norm.weight`,
`layers.*.ffn_norm.weight`) stayed at exactly 1.0 across all 17K
steps of the 2B production run, and likewise across all 4K steps of
the 20B production run.

Linear-layer weights init at much smaller scales (0.005-0.02), so
their bf16 ulp is fine-grained enough that updates accumulate; only
the norm layers were affected.

With `dtype=float32` the master copy is fp32 and FSDP
MixedPrecisionPolicy(param_dtype=bf16, reduce_dtype=fp32) still
keeps forward/backward and reduce in their efficient dtypes — only
the optimizer step is in fp32, where sub-ulp updates can accumulate.

Memory cost: ~1 GB extra master at 2B, ~10 GB at 20B (well within
budget at production scale).

Existing checkpoints inherit the bug; resuming with the new default
will start accumulating real norm updates from the resume step.

+++ torchtitan/experiments/ezpz/agpt/config_registry.py
@@ -66,7 +66,14 @@ def agpt(
     local_batch_size: int = 1,
     activation_checkpoint_mode: Literal["none", "full"] = "full",
     seq_len: int = 8192,
-    dtype: Literal["bfloat16", "float32"] = "bfloat16",
+    # IMPORTANT: bfloat16 master weights silently freeze RMSNorm.weight.
+    # Norm weights init to 1.0 (bf16 ulp = 7.8e-3); per-step updates
+    # ~1.6e-5 round to zero forever, so the model's normalization layers
+    # never train. FSDP MixedPrecisionPolicy keeps the bf16 cast for
+    # forward/backward; reduce stays fp32 — the fp32 master copy is
+    # what enables sub-ulp accumulation.
+    # See docs/guides/known-bugs/training-dtype-bf16-norm-freeze.md.
+    dtype: Literal["bfloat16", "float32"] = "float32",
     compile: bool = True,
     fsdp_reshard_after_forward: Literal["default", "always", "never"] = "default",
     tensor_parallel_degree: int = 1,
